Move from &[u8] to &str as internal representation
It is too useful to have strs, so let's validate once at the start, and
then we can work with str in most of the places.

It makes me really uneasy that slicing strs is more expensive than
slicing a byte sequence due to the multi-byte code point check. We
should be able to avoid it because the lexer would not create tokens
at an invalid boundary. But let's not worry too much about this for now.

Also add a failing test about ascii control characters.