Begin moving towards types in comments
I thought it was pretty neat to abuse quoted strings for type
information, so we could 'select col as "col: String"' and have it be:

 * Valid SQL
 * Somewhat easy to read
 * Contain a type signature

However, that doesn't always work. I want to be able to use structs as
arguments, and that means that I will also need to be able to put type
annotations on parameters.

So let's move the type info into comments, just like it was done in
Python before it gained first-class type annotations. SQL does support
inline comments with /* */ syntax, so add that to the lexer. To avoid
having to deal with the closing part, split comments in a "marker" token
that begins or ends a comment, and a "content" comment that contains
what's inside the comment.