Work towards type annotation comments
This modifies the parser to be able to parse comments that start with a
colon as type annotations. It doesn't yet work in all cases though,
inner annotations inside parens are ignored. I need to make the parser
fully recursive.

As part of this, change the "Type" type to hold a bit more information
about its source spans. Previously it would forget this. That's not a
big issue really, but for pretty-printing, it is nicer if we have the
full original span so we can print it unmodified.

The binary now compiles, but the tests do not. I will fix that in a
follow-up commit.