Parse inside brackets
This removes the special case for being inside brackets. Instead we put
the brackets on a stack so we can make sure they are balanced.

Now that I'm writing this, I wonder ... would we need to ensure this at
all? I could just ignore the brackets entirely, I think. With the new
annotation syntax, I just treat every comment as an annotation. If you
annotate something in e.g. a subquery, that is invalid, but you would
figure that out at runtime, I think that's acceptable.

So maybe the bracket matching is actually a useless gimmick, but it was
fun to implement nonetheless.

Also, now that we put things in a vec, instead of making recursive
calls, there is no longer a risk of overflowing the call stack. And if
you exhaust memory with brackets, well, that's a case of "doctor it
hurts when I press here" "then don't press there". We're not trying to
deal with adversarial inputs here.