Add Nix-based development environment
This is the first time that I'm really trying out flakes, previously I
used used stdenv.buildEnv, which has served me well for a long time --
and it still does. But Nix 2.3 gets quite old now, and it looks like the
ecosystem is moving to flakes, so I should probably adapt.

Instead of Nix 2.3

    nix run --command $SHELL

to enter a development environment, now we need

    nix \
      --extra-experimental-features nix-command \
      --extra-experimental-features flakes \
      develop --command $SHELL

and from then on it should work.