Commit generated output to the repository
It's useful for me to see the diff, if I change the codegen. It might be
useful for interested users to see exactly what kind of code is
generated from a particular input, before they jump in. And if I get
around to doing CI, it would confirm that the generated code works.