Make statement lookup a bit briefer
I don't want to make it all-in code golf as short as possible. But I do
want to have the repeated code be as little of a distraction from the
query as possible. I think this is a fair middle ground, it's not too
cryptic and about as terse as it gets without using a method (which I
tried but failed to write due to lifetimes).