Pin sqlite crate to v0.26.0
The API has some changes in 0.27.0, that make it more difficult to use
in the way in which I intend to use it. The best way to work around it,
it seems, would be to move the statement out of the hash map, and put it
back after we are done. But that makes the code needlessly more
complicated, and also has a needless runtime overhead. (Though probably
tiny compared to executing the query at all, and if I did really care
about that, anyway, I should also use a vec and not a hash map.) There
doesn't seem to be any advantage to using v0.27.0, so pin to v0.26.0 for
now.

It looks like this new limitation was an unintended side effect of the
changes, so hopefully this can be revised in a future version. See also
https://github.com/stainless-steel/sqlite/pull/44#issuecomment-1221576569.