Generate code for selecting the result
And as I do this, I recognize the fatal flaw in my approach. It also
explains why the AST handling is so messy.

The Rust type signature is insufficient for deciding how to drive the
query. Consider these two queries:

   SELECT max(id) FROM users;
   SELECT id FROM users ORDER BY id desc LIMIT 1;

Both queries should return the same Option<i64> in Rust, returning None
when the table is empty. But the first query always returns a single
row, and its first column is nullable, while the second query may return
0 or 1 rows, but the first column is not nullable. (Assuming that the id
column is not nullable.)

So, we can't infer just from the return type how we should iterate the
results. Back to the drawing board ... After some consideration, I think
I will go for different return type arrows:

    @query select_user_by_email(email: str) -?-> User
    @query select_user_by_id(id: i64) -1-> User
    @query select_all_users() -*-> User

This looks kind of neat but it's annoying to type ... how about we move
the sigil after the arrow?

    @query select_user_by_email(email: str) ->? User
    @query select_user_by_id(id: i64) ->1 User
    @query select_all_users() ->* User

Or should it stick to the type then?

    @query select_user_by_email(email: str) -> ?User
    @query select_user_by_id(id: i64) -> =User
    @query select_all_users() -> *User

No, I think on the arrow it is cleaner. And probably we should not allow
a space in between then, to make it clear that it's a property of the
arrow. Or really, they are three different arrows. Keywords are also
still an option, those are more self-documenting.

    @query select_user_by_email(email: str) -> maybe User
    @query select_user_by_id(id: i64) -> one User
    @query select_all_users() -> many User

I like this one, except that it combines Haskell-style "maybe" with
Rust-style Option<T> (which I still want to use for nullable values).
Some alternatives: "option" ... but then it specifies the type again,
not the cardinality, and "option Option<T>" would be confusing. Or
"maxone", but that feels weird to me. Hmmm, maybe sigils after all?

    @query select_user_by_email(email: str) -> {0-1} User
    @query select_user_by_id(id: i64) -> {1} User
    @query select_all_users() -> {1-} User

Nah, I think I'll go with ->?, ->1, ->*.