Introduce new type hierarchy for result types
This hierarchy is more precise, which is more tedious to work with in
somple places, but on the other hand, it makes many invalid cases
unrepresentable. It will simplify the typecheck in the end. And I
believe it will make code generation cleaner as well.

Nonetheless, this is a big and invasive change, so I am starting as
gradually as is possible, by touching only the result types. This breaks
some tests, but at least they still compile.

Next steps would be to eliminate the old 'Type' and 'TypedIdent'
entirely. But let's do one thing at a time.