Use the string literal address as cache key
I thought I'd pre-compute a hash of the SQL and use that as the index
... but we we don't need to think of an identifier for the cache key
manually. Every 'static string literal already has a unique address that
we could use for this.

Although, now I'm writing this ... if we use a compile-time constant,
then LLVM can constant-fold the hash computation and possibly parts of
the hash table lookup. If we use the address of string literal, then
that address is only known at link time, and even with LTO I think the
address would only be known after the code has already been generated.
Possibly even only at runtime, I don't know if string literals are
affected by relocations. But either way, this does maybe destroy some
potential for optimization ... hmm. And also for other target languages
than Rust, this trick may not be available.

This is probably a bad idea ... but it will do for now. Edit: Let's add
a comment to explain this.