Complete migration of types in the AST
Now the query arguments have a dedicated type, that can switch
between struct or a list of arguments. So tuple args are eliminated
by construction. Many other problems are eliminated by construction.
Entire passes of the previous typecheck phase are now no longer needed.
Yay!

This does not yet update the unit tests, but it does update the goldens.
I will update the unit tests in a follow-up.