Support postfix-? as syntax for option types
I don't like the lowercase option<T>, after using Option<i32> or
Optional[int] in Rust and Python, it just looks wrong to me.

But since I added ->? as the syntax for a query that returns zero or one
rows already, it's not that weird any more to use postfix-? for nullable
/ optional types as well. And "i32?" looks okay to me.

This commit does not yet remove support for "optional<T>", I'll do that
in a follow-up commit.