Rename the project to "Squiller"
I was never very satisfied with "Querybinder", it was boring and too
long. I played around with GPT-3, trying to feed it some prompts to make
it suggest other names. It does reasonably well at suggesting names, but
the things it suggests I either found too boring, or there were already
projects with that name. "Squiller" was not among the GPT-3 suggestions,
but some of its outputs did inspire me. Some other GPT-generated names:

 - querymatic
 - queryable
 - queryfy
 - querify
 - querious
 - queryness
 - queriosity
 - querent
 - queryist
 - queryize
 - queryous
 - sqlmill
 - querymill
 - boilermaker
 - querygen
 - query-gen
 - genquery
 - squealer
 - querymaker
 - querywrangler
 - sqlwrangler

I think "Squealer" crossed my mind before too. Anyhow, I think
"Squiller" is an improvement over "Querybinder", it's shorter, less
dull, more distinct. Also no risk of people capitalizing it
"QueryBinder". It is a bit difficult to type though.

This does bring us back very close to my original name, "Sqoiler". Which
is maybe nicer because it has the "oiler" of "boilerplate" in it, but
which is also even more impossible to type.