Add a hand-written CLI parser
I have done this before, so the args iterator is adapted from Tako. This
does not yet remove Clap or adapt the main program to call this parser,
it only adds the parser and tests. The intention is to remove Clap
though, because it's an unreasonably large part of the binary; overkill
for my purpose. Also, expressing what I want with Clap is actually kind
of difficult. A bespoke parser will do nicely.