Allow multiple statements in the AST
I want to be able to write queries that consist of multiple statements,
so step one is to modify the AST to hold multiple statements, instead of
a single one. This does not yet modify the parser, and the code
generators take a shortcut and handle only a single statement for now.