Remove statement mode/type from AST
On the one hand, yes it is part of the input syntax. On the other hand,
for code generation we only care about the statements, the field is
never read after we construct the document AST, it is really only a
parsing directive. To make invalid states (StatementMode::Single, but
having multiple statements in the vec) unrepresentable, let's remove the
mode from the Annotation type, and return it out of band.

This does make the AST lossy, since you can write a single-statement
function with @begin and @end markers, and this way we cannot
distinguish that from an @query marker in the AST. But I think that is
fine, so far I do not care about being able to print at that level, the
debug printer also does not print verbatim.