Write preamble for Python code
This code is inspired by the Transaction and Connection class that I
wrote for Hanson, my prediction market app. It is working well there
so far, I am donating this code here as boilerplate for Squiller to
include. Currently it has a few things hard-coded to my needs, like the
serializable isolation level, but that can all be generalized or made
configurable later.

For now, this only generates the preamble and header comment, no
actually useful code yet.