Add fuzz target that covers code generation
This fuzzer now covers the full pipeline: parsing, typechecking, and
generating code for the target. We can fuzz all targets at once with a
single fuzzer by sampling the target iself from the fuzzer.

So far it did catch a panic in the rust-sqlite target; I didn't look
into the details yet.